Extract Standard Surface BSDF into metashade.mtlx.standard_surface
Move the BSDF implementation from the test file into a production
module with a generate() entry point typed against GlslGeneratorContext.
The hand-written nodegraph moves to metashade/mtlx/libraries/ so it
ships as package data alongside the module and is loadable via the
standard MaterialX loadLibraries API.

The test becomes a thin wrapper that delegates to generate().
